requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai
python-dotenv
opensearch-py
ragas
langchain-google-genai
pandas
